Temporary Commit
Some aspects of this changeset (specifically, trying to remove the
extra `/nhc` when installing to a `prefix` that already has `/nhc` in
it, like `/opt/nhc/1.5` or `~/.local/nhc`) are not working properly
and causing me fits.  Someday I'd like to make this work, so I'm going
to commit my work in progress, revert, and then commit a working
solution, albeit with redundancy remaining. :-(