nhc & build:  Flesh out autotools prefix support
Still working with @papamoose on user support, we found a couple more
issues.  For one, as you can see in the commented-out blocks in `nhc`
that disabled auto-creation of installation directories, I don't think
`nhc` -- that is, the script itself -- should be creating directories
for itself at runtime.  Missing mount points, sure, but not for things
like `/etc/nhc` or `/var/log`.

So instead, what I did was to make sure that the assumptions and
fallbacks that `nhc` itself uses internally are the same ones being
created/managed by the GNU Autotools infrastucture.  In looking into
this, I noticed a typo in both the specfile and the `Makefile.am`
Automake file, so I corrected that (`/var/lib/nhc` --> `/var/log`).

I also tweaked a couple of debugging messages to be more helpful and
removed a complex expansion that broke Bash 4.x, but the main goal of
this commit is to get the `make install` process back in sync with the
driver script.